+++ hata/discord/soundboard/soundboard_sounds_event/tests/test__SoundboardSoundsEvent__data.py
@@ -34,7 +34,7 @@ def test__SoundboardSoundsEvent__from_data__1():
     """
     Tests whether ``SoundboardSoundsEvent.from_data`` works as intended.
     
-    Case: all fields given.
+    Case: caching new.
     """
     guild_id = 202305280010
     sound_id_0 = 202305280011
@@ -54,6 +54,33 @@ def test__SoundboardSoundsEvent__from_data__1():
     
     vampytest.assert_true(guild.soundboard_sounds_cached)
     vampytest.assert_eq(guild.soundboard_sounds, {sound_id_0: sound_0, sound_id_1: sound_1})
+
+
+def test__SoundboardSoundsEvent__from_data__2():
+    """
+    Tests whether ``SoundboardSoundsEvent.from_data`` works as intended.
+    
+    Case: caching clearing.
+    """
+    guild_id = 202305290005
+    sound_id_0 = 202305290006
+    sound_id_1 = 202305290007
+    sound_0 = SoundboardSound.precreate(sound_id_0, guild_id = guild_id)
+    sound_1 = SoundboardSound.precreate(sound_id_1, guild_id = guild_id)
+    sounds = [sound_1]
+    
+    guild = Guild.precreate(guild_id)
+    guild.soundboard_sounds = {sound_id_0: sound_0}
+    
+    data = {
+        'guild_id': str(guild_id),
+        'soundboard_sounds': [sound.to_data(defaults = True, include_internals = True) for sound in sounds],
+    }
+    
+    SoundboardSoundsEvent.from_data(data)
+    
+    vampytest.assert_true(guild.soundboard_sounds_cached)
+    vampytest.assert_eq(guild.soundboard_sounds, {sound_id_1: sound_1})
 
 
 def test__SoundboardSoundsEvent__to_data__0():